Fix error when using lowecase LOG_LEVEL (#491)
* Apply .gitattributes normalization

* Force all files to be LF

* Better logs when docker socket is unavailable

* Update `run` to be LF

* Force files to be LF

* Fix logger not handling lowercase entries

---------

+++ app/utils.sh
@@ -74,6 +74,13 @@ initialize_nautical() {
     if [ ! -f "/usr/local/bin/nautical" ]; then
         logThis "Installing nautical backup script..." "DEBUG" "init"
         # Allows the nautical backup script to be run using `bash nautical`
+        
+        # Check if file exists
+        if [ ! -f "/app/nautical.sh" ]; then
+            logThis "Nautical backup script not found: /app/nautical.sh" "ERROR" "init"
+            exit 1
+        fi
+        
         ln -s /app/nautical.sh /usr/local/bin/nautical
         chmod +x /usr/local/bin/nautical
         logThis "ln -s /app/backup.sh /usr/local/bin/nautical" "TRACE" "init"
